refactor : 댓글 기능 인가 적용

+++ src/main/java/com/study/proudcat/domain/comment/service/CommentService.java
@@ -5,12 +5,16 @@ import com.study.proudcat.domain.comment.entity.Comment;
 import com.study.proudcat.domain.comment.repository.CommentRepository;
 import com.study.proudcat.domain.post.entity.Post;
 import com.study.proudcat.domain.post.repository.PostRepository;
+import com.study.proudcat.domain.user.entity.User;
+import com.study.proudcat.domain.user.repository.UserRepository;
 import com.study.proudcat.infra.exception.ErrorCode;
 import com.study.proudcat.infra.exception.RestApiException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Objects;
 
 @Slf4j
 @Service
@@ -19,23 +23,33 @@ public class CommentService {
 
     private final CommentRepository commentRepository;
     private final PostRepository postRepository;
+    private final UserRepository userRepository;
 
     @Transactional
-    public void writeComment(Long postId, CommentRequest request) {
+    public void writeComment(Long postId, CommentRequest request, Long userId) {
         log.info("Comment service writeComment run..");
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new RestApiException(ErrorCode.NO_TARGET));
-
-        Comment comment = Comment.of(request, post);
-        log.info("comment : {}", comment);
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new RestApiException(ErrorCode.NO_TARGET));
+        Comment comment = Comment.builder()
+                .content(request.getContent())
+                .post(post)
+                .user(user)
+                .build();
 
         commentRepository.save(comment);
     }
 
     @Transactional
-    public void deleteComment(Long postId, Long commentId) {
+    public void deleteComment(Long postId, Long commentId, Long userId) {
         log.info("Comment service deleteComment run..");
         Comment comment = errorCheckComment(postId, commentId);
+
+        if (!Objects.equals(comment.getUser().getId(), userId)) {
+            throw new RestApiException(ErrorCode.NOT_PROPER_USER);
+        }
+
         commentRepository.delete(comment);
     }
